requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
Automat==24.8.1
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==44.0.2
cssselect==1.2.0
ctransformers==0.2.27
defusedxml==0.7.1
Deprecated==1.2.18
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
durationpy==0.9
elastic-transport==8.17.1
elasticsearch==8.18.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.17.0
flatbuffers==25.2.10
frozenlist==1.6.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.69.1
gpt4all==2.8.2
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==32.0.1
langchain==0.3.20
langchain-core==0.3.41
langchain-text-splitters==0.3.6
langsmith==0.3.12
llama_cpp_python==0.3.8
lxml==5.3.1
lxml_html_clean==0.4.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
networkx==3.4.2
numpy==2.2.3
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.68.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
parsel==1.10.0
pillow==11.1.0
posthog==3.19.0
propcache==0.3.1
Protego==0.4.0
protobuf==5.29.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-extra-types==2.10.3
pydantic-settings==2.9.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
PyJWT==2.10.1
PyNaCl==1.5.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pywebcopy==7.0.2
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.14.3
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
Scrapy==2.12.0
sentence-transformers==3.4.1
service-identity==24.2.0
setuptools==75.8.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.0
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0
Twisted==24.11.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.34.0
w3lib==2.3.1
watchfiles==1.0.4
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
